[ansatz] Added more optimizers and started cleanup.
Now we have two ways to fulfill the constraints in the optimization:
- Use an optimizer that can deal with the constraints
- Add a L1/L2 penalty to the loss that enforces the constraints

Todo:
- Unit test.
